refactor: rename SwiftElement.py -> Elements.py, avoiding a module/class name collision
Discovered while type-hinting: a file named SwiftElement.py containing
a class also named SwiftElement, re-exported at the package's top
level, confused mypy into resolving `SwiftElement` as the submodule
in several contexts instead of the class -- ~20 of the 33 mypy
findings from that pass traced back to this one root cause. Doesn't
affect runtime (Python's own import resolution handles it fine,
confirmed by the full test suite both before and after), but is
easy to just fix outright rather than carry as permanent static-
analysis noise.

Renamed src/swift/SwiftElement.py -> src/swift/Elements.py and its
dedicated test file tests/test_swift_element.py ->
tests/test_elements.py (test_assembly_handle.py keeps its name --
it only imports Slider incidentally, isn't the Elements module's own
test file). Updated every direct `swift.SwiftElement` module-path
reference: src/swift/__init__.py, tests/test_assembly_handle.py,
examples/collision.py, and three docs/source/*.rst files
(api.rst's automodule, internals.rst's cross-reference, intro.rst's
code example). Swift.py needed no change -- it already imports
SwiftElement from the top-level `swift` package, not the submodule
directly.

Verified: full test suite passes, examples/collision.py runs
correctly headless, sphinx-build -W (warnings as errors) succeeds
with zero warnings and correctly picks up the renamed module.
