+++ docs/source/intro.rst
@@ -268,7 +268,7 @@ The distance between the two shapes is displayed in a label, and when the sphere
     import spatialgeometry as sg
     from spatialmath import SE3
     from swift import Swift
-    from swift.SwiftElement import Slider, Label
+    from swift.Elements import Slider, Label
 
     env = Swift()
     env.launch(realtime=True, ground_opacity=0.1)
